finish equations

+++ lib/models/equipment_ct_mdoel.dart
@@ -1,5 +1,5 @@
 class EQuipmentCT {
-  int dumperTrucker = 0;
-  int actros6Wheels = 0;
-  int actros6WheelsPh = 0;
+  double dumperTrucker = 0;
+  double actros6Wheels = 0;
+  double actros6WheelsPh = 0;
 }
